+++ include/compile/info.hpp
@@ -1,0 +1,25 @@
+
+// --- Kartonagnick/compile                                           [info.hpp]
+// [2025-11-06][18:10:00] 001 Kartonagnick PRE
+//   --- CastleOfDreams/compile                                       [info.hpp]
+//   [2022-08m-17][19:00:00] 001 Kartonagnick
+//     --- Kartonagnick/buid_info (closed)                      [build_info.hpp]
+//     [2021-07-02][14:00:00] 003 Idrisov Denis R.
+//     [2021-06-18][22:50:00] 002 Idrisov Denis R.
+//     [2021-06-13][02:00:00] 001 Idrisov Denis R.
+
+//==============================================================================
+//==============================================================================
+
+#pragma once
+#ifndef dCOMPILE_INFO_USED_ 
+#define dCOMPILE_INFO_USED_ 1
+
+#include "message.hpp"
+#include "private/cplusplus.hpp"
+#include "private/select.hpp"
+#include "private/show.hpp"
+
+//==============================================================================
+//==============================================================================
+#endif // dCOMPILE_INFO_USED_